backend/python-face-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
insightface>=0.7.3
onnxruntime>=1.17.0
scikit-learn>=1.3.0

